% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_eigengenes.R
\name{extract_eigengenes}
\alias{extract_eigengenes}
\title{Extract module eigengenes from an expression set}
\usage{
extract_eigengenes(
  eset,
  signatures,
  mod_fdata = NULL,
  key = NULL,
  min_size = 3,
  method = c("multiple", "single"),
  name_cleanup = TRUE
)
}
\arguments{
\item{eset}{an \code{ExpressionSet}}

\item{signatures}{named list of character vectors of features}

\item{mod_fdata}{optional feature-level metadata for modules}

\item{key}{if \code{NULL} match signatures to \code{featureNames(eset)},
otherwise use \code{fData(eset)[, key]}}

\item{min_size}{minimum number of overlapping features required per signature}

\item{method}{eigengene extraction method; use \code{"single"} to run modules
independently and \code{"multiple"} for one joint call}

\item{name_cleanup}{logical; remove leading \code{"ME"} from module names}
}
\value{
an \code{ExpressionSet} with eigengenes as features and input samples as
columns
}
\description{
Extract module eigengenes from an expression set
}
