% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_pc.R
\name{extract_pc}
\alias{extract_pc}
\title{Extract a principal component score for a signature}
\usage{
extract_pc(eset, sig, i = 1)
}
\arguments{
\item{eset}{an \code{ExpressionSet}}

\item{sig}{character vector of feature names in \code{eset}}

\item{i}{index of the principal component to extract}
}
\value{
named numeric vector of PC scores (one per sample)
}
\description{
Extract a principal component score for a signature
}
