% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{signature}
\alias{signature}
\title{Example Signature}
\format{
A named list containing one character vector of gene identifiers.
}
\usage{
data(signature)
}
\description{
Example gene signature as a named list.
}
\keyword{datasets}
