% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spc_heatmaps.R
\name{spc_heatmap_sig}
\alias{spc_heatmap_sig}
\title{Generate Signature-Only SPC Heatmap}
\usage{
spc_heatmap_sig(
  eset,
  spc_out,
  col_ha = NULL,
  min_sigsize = 3,
  name = "expression",
  ...
)
}
\arguments{
\item{eset}{an expression set}

\item{spc_out}{output list from \code{signature_projection_contributors()}}

\item{col_ha}{a ComplexHeatmap::heatmapAnnotation object with columns' (i.e., samples') annotation}

\item{min_sigsize}{minimum number of signature genes required to plot signature-only heatmap}

\item{name}{name for the heatmap}

\item{...}{additional parameters to pass to \code{ComplexHeatmap::Heatmap}}
}
\value{
a ComplexHeatmap object
}
\description{
Build the signature-only heatmap from a
\code{signature_projection_contributors()} result.
}
