% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{toy_eset}
\alias{toy_eset}
\title{Toy ExpressionSet}
\format{
An \code{ExpressionSet} object with genes in rows and samples in columns.
}
\usage{
data(toy_eset)
}
\description{
Small toy expression matrix packaged as an \code{ExpressionSet}.
}
\keyword{datasets}
