% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{toy_signature}
\alias{toy_signature}
\title{Toy Signature}
\format{
A character vector of gene identifiers.
}
\usage{
data(toy_signature)
}
\description{
Small toy gene signature as a character vector of gene identifiers.
}
\keyword{datasets}
