% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signature_projection_contributors.R
\name{signature_projection_contributors}
\alias{signature_projection_contributors}
\title{This function generates an annotated heatmap that shows the contributions of genes to the given signature score}
\usage{
signature_projection_contributors(
  eset,
  signature,
  sig_score = NULL,
  cor_method = c("pearson", "spearman", "kendall"),
  col_ha = NULL,
  min_sigsize = 3,
  method = c("GSVA"),
  name = "expression",
  gsea = FALSE,
  ...
)
}
\arguments{
\item{eset}{an expression set}

\item{signature}{a list of signatures (at least 1)}

\item{sig_score}{an aggregate signature score}

\item{cor_method}{correlation method used by \code{psych::corr.test}}

\item{col_ha}{a ComplexHeatmap::heatmapAnnotation object with columns' (i.e., samples') annotation}

\item{min_sigsize}{minimum number of signature genes required to plot signature-only heatmap}

\item{method}{how to compute the aggregate score (only GSVA at the moment)}

\item{name}{name for the heatmap}

\item{gsea}{logical, whether to use weighted KS statistic}

\item{...}{additional parameters to pass to the method}
}
\value{
a list of a dataframe and two heatmaps
}
\description{
This function generates an annotated heatmap that shows the contributions of genes to the given signature score
}
